playing with print methods

+++ R/ontology_topgo.R
@@ -43,7 +43,7 @@ simple_topgo <- function(sig_genes, goid_map = "id2go.map", go_db = NULL,
   ## set up the GOdata object like this: mf_GOdata = new("topGOdata",
   ## description = "something", ontology = "BP", allGenes = entire_geneList,
   ## geneSel = topDiffGenes, annot = annFUN.gene2GO, gene2GO = geneID2GO, nodeSize = 2)
-  if (isTRUE(overwrite) & file.exists(goid_map)) {
+  if (isTRUE(overwrite) && file.exists(goid_map)) {
     removed <- file.remove(goid_map)
   }
   gomap_info <- make_id2gomap(goid_map = goid_map, go_db = go_db, overwrite = overwrite)
